Rename to TerminalGuiFluentTesting
