+++ Terminal.Gui/Terminal.Gui.csproj
@@ -86,7 +86,7 @@
         <InternalsVisibleTo Include="StressTests" />
         <InternalsVisibleTo Include="IntegrationTests" />
         <InternalsVisibleTo Include="TerminalGuiDesigner" />
-        <InternalsVisibleTo Include="TerminalGuiFluentAssertions" />
+        <InternalsVisibleTo Include="TerminalGuiFluentTesting" />
         <InternalsVisibleTo Include="DynamicProxyGenAssembly2" />
     </ItemGroup>
     <!-- =================================================================== -->
