+++ TerminalGuiFluentAssertions/FakeNetInput.cs
@@ -1,0 +1,6 @@
+﻿using Terminal.Gui;
+
+namespace TerminalGuiFluentTesting;
+
+internal class FakeNetInput (CancellationToken hardStopToken) : FakeInput<ConsoleKeyInfo> (hardStopToken), INetInput
+{ }
