+++ TerminalGuiFluentAssertions/FakeWindowsInput.cs
@@ -1,0 +1,6 @@
+﻿using Terminal.Gui;
+
+namespace TerminalGuiFluentTesting;
+
+internal class FakeWindowsInput (CancellationToken hardStopToken) : FakeInput<WindowsConsole.InputRecord> (hardStopToken), IWindowsInput
+{ }
